Springboard24_10/backend/requirements.txt
fastapi
pydantic
motor
pymongo
python-dotenv
uvicorn
email-validator
pandas
openpyxl
openai
transformers
torch
numpy
scipy
textblob
vaderSentiment
nltk
livekit-api